Move entry into its own module
Hmm, Logger doesn't depend on log.
