+++ src/accountant_stub.rs
@@ -8,7 +8,7 @@ use bincode::{deserialize, serialize};
 use transaction::Transaction;
 use signature::{PublicKey, Signature};
 use hash::Sha256Hash;
-use log::Entry;
+use entry::Entry;
 use ring::signature::Ed25519KeyPair;
 use accountant_skel::{Request, Response};
