+++ src/logger.rs
@@ -8,7 +8,7 @@
 use std::sync::mpsc::{Receiver, SyncSender, TryRecvError};
 use std::time::{Duration, Instant};
 use hash::Sha256Hash;
-use log::{create_entry_mut, Entry};
+use entry::{create_entry_mut, Entry};
 use event::Event;
 use serde::Serialize;
 use std::fmt::Debug;
